Fixed physical memory management
We do not have identity mapping when starting anymore. This was done by UEFI, but only loader has this now. So we have to allocate physical and virtual pages and map the physical page to an allocated virtual one.

In other words: proper virtual memory management seems to be here? \o/