Split arch/generic, kernel now ELF
Moved a lot of arch-agnostic files out of the arch/amd64 directory and
adapted the build system to build the arch-agnostic sources to a static
library, which is then linked in the arch-specific kernel.

Kernel is now linked to an ELF binary file and this file is now loaded
by the EFI loader, there is no flat binary file anymore. The loader
supports ELF binaries now (and only those), but enforces the address
documented in src/kernel/arch/amd64/vm.h for the kernel binary. No error
checking is done!

fbconsole learnt a new trick: '\r', this removed the dependency on the
fbconsole data structure internals for INIT_STEP().

One last thing. Assembly files are now assembled with gcc/gas and not
nasm anymore. I was not able to tell nasm to generate 64bit relocations
in the object files, resulting in not being able to link them.