WIP: do not deschedule on syscalls
Only processes descheduled by a syscall can safely be scheduled by
handling a syscall, as registers are modified while returning from
syscall to userspace. Easiest fix is to not deschedule processes on
syscalls, for which a part is implemented in this commit but more needs
to be done (like yield()).