sys/message_passing.h: include sys/types.h to be standalone
.. at least if not building the kernel rn. This is a HACK, instead
kernel and userspace should just have the same sys/types.h but since
newlib has A LOT in there and I'm not currently in the mood to do
another weird workaround as with errno.h (see sys/errno-defs.h), this is
hacky for now. Deal with it.

Merge requests welcome though :)