kernel/elf.c: fix bug with unaligned segments
Even with segment alignment set to a full page, linkers may produce
binaries where the virtual address (and file offset) are not aligned to
a page. This apparently only leads to segments not overlapping on that
alignment.

LF OS has code in place to deal with that, dealing with that first page
being incomplete and only loading data until the next page boundary.
Just one problem: the logic was wrong and not all data was copied.

Fix that logic, unbreaking my test program and I think that code is
actually a bit nicer now, too.

Reference thread:
https://gotosocial-dev.svc.0x0a.network/@littlefox/statuses/01HDWA8YBQVG8F5QCWZTP7W4C6