loader: (probably) detect if UART is present
The loader should now detect if an UART is actually present, only
printing to it in that case instead of going into an infinite "wait for
transmit buffer to be empty".

Also changes the check for "EDK II" firmware to a debian specific string
for disabling logging to UART - I hate I can't just query that.