qr: correct endianness for EC coding
Previously, data was prepared in the bit-order for QR codes and that was passed to the error correction coding, leading to endianness problems. Now preparing the data properly and change the bit order while placing the bits into the matrics.

This is only correct for little-endian though, but that should be fine (TM).

Also removes trailing whitespaces left in rs.c