disable legacy PIT and use HPET instead
Some computers (like the Tuxedo InfinityBook Pro 14 Gen7 (MK1) that I
acquired especially because it had an interesting LF OS boot failure (oh
goth, how far am I gone now - getting a computer because it fails in an
interesting way?!)) apparently just don't have a working PIT anymore -
so let's just get rid of that piece of legacy tech and use the HPET for
now.

Why HPET and not LAPIC or something? Well, we have an HPET driver
already and require that piece of hardware - might as well make more use
of it.